refactor: core typing changes

+++ lib/include/pico_ssd1306/communication.h
@@ -3,7 +3,7 @@
 #include "pico_ssd1306/types.h"
 
 void _write_to_reg(SSD1306_I2C i2c_c, uint8_t reg_address, uint8_t buffer[],
-                   int bufflen);
+                   uint16_t bufflen);
 void _read_from_reg(SSD1306_I2C i2c_c, uint8_t reg, int bytes_to_read,
                     uint8_t* buff);
 void _send_commands(SSD1306_I2C i2c_c, uint8_t cmds[], int commands_count);
